more controllers, upload almost complete in backend, santi needs to work

+++ pawgram/webapp/src/main/java/ar/edu/itba/pawgram/webapp/auth/PawgramUserService.java
@@ -32,7 +32,7 @@ public class PawgramUserService implements SecurityUserService {
     }
 
     @Override
-    public User changePassword(int userId, String password) {
+    public User changePassword(long userId, String password) {
         return userService.changePassword(userId,password);
     }
 }
